refactor: simplify test suite — remove duplicates, fix silent skips
- Remove BeveledShapeForTypeTest.java (8 tests duplicated by Deep+Comprehensive,
  with incorrect headless guards causing 8 unnecessary CI skips)
- Remove TypeManagerTest.java (7 tests subsumed by DeepTest+ExtendedTest)
- Remove IdentifierNameGeneratorExtendedTest.java (9/14 tests duplicated);
  merge 5 unique tests into ComprehensiveTest
- Fix 3 silent skips in ComprehensiveTest: replace if-guards with assertions
  so tests fail explicitly instead of passing vacuously
- Remove 3 duplicate getEnumConstantField tests from TypeManagerDeepTest
  (already covered by TypeManagerExtendedTest)

Result: 12 files, 240 tests (237 pass, 3 legitimately headless-skipped),
2708 lines. Net reduction of 233 lines and 27 duplicate/skipped tests.
